WIP: Refactoring node app
