+++ api/dbOperations/database.js
@@ -1,16 +1,4 @@
-import { DatabaseOps } from "./dbops.js";
-
-const dbConfig = {
-    type: process.env.DB_PROTOCOL,
-    connectionLimit: 10,
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DATABASE,
-    multipleStatements: true
-}
-
-const dbOps = new DatabaseOps(dbConfig)
+import dbOps from "./dbConnection.js"
 
 export async function getUserByUsername(username) {
     const query = `SELECT * FROM users WHERE username = ?`
@@ -109,114 +97,6 @@ export async function getUsers(userId) {
     const rows = await dbOps.select(query, [userId])
 
     return rows
-}
-
-export async function getTasks(userId) {
-    const query = `SELECT id, prompt, short_name
-                FROM tasks
-                WHERE investigator = ?`
-    const rows = await dbOps.select(query, [userId])
-    
-    return rows
-}
-
-export async function createTask(userId, short_name, prompt) {
-    const query = `INSERT INTO tasks (short_name, prompt, investigator) VALUES (?, ?, ?);`
-    try {
-        const results = await dbOps.executeWithResults(query, [short_name, prompt, userId])
-        return results.insertId
-    } catch (err) {
-        throw(err)
-    }
-
-}
-
-export async function updateTask(userId, taskId, short_name, prompt) {
-    const query = `UPDATE tasks
-                SET short_name = ?,
-                    prompt = ?
-                WHERE investigator = ?
-                    AND id = ?`
-    try {
-        await dbOps.execute(query, [short_name, prompt, userId, taskId])
-        return true
-    } catch (err) {
-        throw (err)
-    }
-
-}
-
-export async function deleteTask(userId, taskId) {
-    const query = `DELETE FROM tasks WHERE tasks.investigator = ? AND tasks.id = ?;`
-    try {
-        await dbOps.execute(query, [userId, taskId])
-        return true
-    } catch (err) {
-        throw (err)
-    }
-}
-
-// export async function getTask(userId, taskId) {
-//     const query = `SELECT id, prompt, short_name
-//                 FROM tasks
-//                 WHERE investigator = ?
-//                 AND id = ?`
-//     const rows = await dbOps.select(query, [userId, taskId])
-
-//     return rows
-// }
-export async function getTaskTable(userId) {
-    // TODO: I think this can probably be clean up a little
-    // getting image counts should only require the image_tags table,
-    // and there maybe other redundant joins. For now it works!
-    const query = `WITH image_count_table AS (
-  SELECT tasks.id AS task_id, COUNT(DISTINCT images.id) AS image_count
-  FROM tasks
-  LEFT JOIN task_tags ON task_tags.task_id = tasks.id
-  LEFT JOIN image_tags ON image_tags.tag_id = task_tags.tag_id
-  LEFT JOIN images ON images.id = image_tags.image_id
-  WHERE tasks.investigator = ?
-  GROUP BY tasks.id
-),
-observer_count_table AS (
-  SELECT tasks.id AS task_id, COUNT(DISTINCT observers.user_id) AS observer_count
-  FROM tasks
-  LEFT JOIN observers ON observers.task_id = tasks.id
-  WHERE tasks.investigator = ?
-  GROUP BY tasks.id
-),
-progress_table AS (
-	SELECT hotornot.user_id, COUNT(DISTINCT hotornot.image_id) AS graded_images_count, hotornot.task_id,
-    total_images.total_images,
-    (COUNT(DISTINCT hotornot.image_id) / total_images.total_images) AS progress_percentage
-  FROM hotornot
-  JOIN (
-    SELECT tasks.id as tt_id, COUNT(DISTINCT images.id) AS total_images
-    FROM tasks
-    LEFT JOIN task_tags ON task_tags.task_id = tasks.id
-    LEFT JOIN image_tags ON image_tags.tag_id = task_tags.tag_id
-    LEFT JOIN images ON images.id = image_tags.image_id
-    GROUP BY tasks.id
-  ) AS total_images ON total_images.tt_id = hotornot.task_id
-  GROUP BY hotornot.task_id, hotornot.user_id
-),
-overall AS (
-	SELECT progress_table.task_id AS task_id, SUM(progress_table.progress_percentage)/COUNT(progress_table.progress_percentage) AS overall_progress
-  FROM progress_table
-  GROUP BY progress_table.task_id
-)
-SELECT tasks.id, tasks.short_name, tasks.prompt,
-       COALESCE(image_count_table.image_count, 0) AS image_count,
-       COALESCE(observer_count_table.observer_count, 0) AS observer_count,
-       COALESCE(overall.overall_progress, 0) AS progress
-FROM tasks
-LEFT JOIN image_count_table ON tasks.id = image_count_table.task_id
-LEFT JOIN observer_count_table ON tasks.id = observer_count_table.task_id
-LEFT JOIN overall ON tasks.id = overall.task_id;`
-
-    const rows = await dbOps.select(query, [userId, userId])
-    return rows
-
 }
 
 export async function getData(userId, taskId) {
